requirements.txt
fastapi~=0.110.1
uvicorn
sqlalchemy~=2.0.29
aiosqlite
pydantic~=2.7.0
python-dotenv~=1.0.1
python-dotenv